requirements.txt
# Tested Working Combination - Latest Compatible Versions
# August 2025 - No dependency conflicts

# Core FastAPI dependencies
fastapi==0.112.0
uvicorn[standard]==0.30.5
python-multipart==0.0.9

# Pydantic v2 - Stable version
pydantic==2.8.2

# LangChain ecosystem - Let pip resolve compatible versions
langchain>=0.2.0,<0.3.0
langchain-community>=0.2.0,<0.3.0
langchain-google-genai>=1.0.0
langchain-core>=0.2.0,<0.3.0

# Vector database - Latest stable
chromadb>=0.5.0

# AI frameworks - Compatible versions
crewai>=0.55.0
google-generativeai>=0.7.0

# Utilities
python-dotenv==1.0.1
PyPDF2==3.0.1
tiktoken>=0.7.0

# OpenAI - Latest
openai>=1.40.0

# Additional dependencies that help resolve conflicts
httpx>=0.27.0
requests>=2.32.0
typing-extensions>=4.8.0
pydantic-settings==2.4.0